Address C130/C131 review follow-ups
