+++ crates/shuck-linter/src/rules/correctness/shell_quoting_reuse.rs
@@ -3,7 +3,9 @@ use shuck_ast::{Command, DeclOperand, Span, Word};
 use shuck_semantic::{BindingId, BindingKind, Reference, ReferenceKind};
 
 use crate::facts::CommandId;
-use crate::{Checker, ExpansionContext, WordFactContext};
+use crate::{
+    Checker, ExpansionContext, WordFactContext, word_shell_quoting_literal_run_span_in_source,
+};
 
 pub(crate) struct ShellQuotingReuseAnalysis {
     pub assignment_spans: Vec<Span>,
@@ -193,12 +195,7 @@ fn plain_scalar_reference_bindings(
     references: &[Reference],
     reference_indices: &[usize],
 ) -> Vec<BindingId> {
-    let Some(fact) = checker
-        .facts()
-        .word_facts()
-        .iter()
-        .find(|fact| fact.span() == word_span)
-    else {
+    let Some(fact) = checker.facts().any_word_fact(word_span) else {
         return Vec::new();
     };
 
@@ -431,253 +428,8 @@ fn repeated_export_assignment_targets(checker: &Checker<'_>) -> FxHashSet<String
 }
 
 fn assignment_value_report_span(binding: ScalarBinding<'_>, checker: &Checker<'_>) -> Span {
-    let source = checker.source();
-    let word = binding.word;
-    let Some(span) = source_shell_quoting_literal_run_span(word, source) else {
-        return word.span;
-    };
-    span
-}
-
-fn source_shell_quoting_literal_run_span(word: &Word, source: &str) -> Option<Span> {
-    let text = word.span.slice(source);
-    let mut cursor = if word.is_fully_double_quoted() && text.starts_with('"') {
-        1
-    } else {
-        0
-    };
-    let limit = if word.is_fully_double_quoted() && text.ends_with('"') {
-        text.len().saturating_sub(1)
-    } else {
-        text.len()
-    };
-    let mut saw_expansion = false;
-    let mut in_single = false;
-    let mut in_double = word.is_fully_double_quoted() && text.starts_with('"');
-    let mut index = cursor;
-
-    while index < limit {
-        let tail = &text[index..limit];
-        let Some(ch) = tail.chars().next() else {
-            break;
-        };
-        if ch == '\'' && !in_double && !text_position_is_escaped(text, index) {
-            in_single = !in_single;
-            index += ch.len_utf8();
-            continue;
-        }
-        if ch == '"' && !in_single && !text_position_is_escaped(text, index) {
-            in_double = !in_double;
-            index += ch.len_utf8();
-            continue;
-        }
-        if !in_single && matches!(ch, '$' | '`') && !text_position_is_escaped(text, index) {
-            saw_expansion = true;
-            if let Some(span) = shell_quoting_segment_span(word, text, cursor, index) {
-                return Some(span);
-            }
-            index += expansion_len(tail);
-            cursor = index;
-            continue;
-        }
-        index += ch.len_utf8();
-    }
-
-    if let Some(span) = shell_quoting_segment_span(word, text, cursor, limit) {
-        return Some(span);
-    }
-    if !saw_expansion && source_text_contains_shell_quoting_literals(&text[..limit]) {
-        return Some(word.span);
-    }
-
-    None
-}
-
-fn expansion_len(text: &str) -> usize {
-    if text.starts_with('`') {
-        return closing_backtick_offset(text).unwrap_or(1);
-    }
-    if !text.starts_with('$') {
-        return 1;
-    }
-
-    if text.starts_with("${") {
-        return braced_expansion_len(text).unwrap_or(2);
-    }
-    if text.starts_with("$(") {
-        return paren_expansion_len(text).unwrap_or(2);
-    }
-
-    let bytes = text.as_bytes();
-    let Some(&next) = bytes.get(1) else {
-        return 1;
-    };
-    if (next as char).is_ascii_alphabetic() || next == b'_' {
-        let mut end = 2usize;
-        while let Some(byte) = bytes.get(end) {
-            let ch = *byte as char;
-            if ch.is_ascii_alphanumeric() || ch == '_' {
-                end += 1;
-                continue;
-            }
-            break;
-        }
-        return end;
-    }
-    if (next as char).is_ascii_digit() || b"@*#?$!-".contains(&next) {
-        return 2;
-    }
-
-    1
-}
-
-fn closing_backtick_offset(text: &str) -> Option<usize> {
-    let mut chars = text.char_indices();
-    chars.next()?;
-    for (offset, ch) in chars {
-        if ch == '`' && !text_position_is_escaped(text, offset) {
-            return Some(offset + 1);
-        }
-    }
-
-    None
-}
-
-fn braced_expansion_len(text: &str) -> Option<usize> {
-    let mut depth = 0usize;
-    for (offset, ch) in text.char_indices() {
-        match ch {
-            '$' if offset == 0 => {}
-            '{' => depth += 1,
-            '}' => {
-                depth = depth.saturating_sub(1);
-                if depth == 0 {
-                    return Some(offset + 1);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    None
-}
-
-fn paren_expansion_len(text: &str) -> Option<usize> {
-    let mut depth = 0usize;
-    for (offset, ch) in text.char_indices() {
-        match ch {
-            '$' if offset == 0 => {}
-            '(' => depth += 1,
-            ')' => {
-                depth = depth.saturating_sub(1);
-                if depth == 0 {
-                    return Some(offset + 1);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    None
-}
-
-fn text_position_is_escaped(text: &str, offset: usize) -> bool {
-    let bytes = text.as_bytes();
-    let mut cursor = offset;
-    let mut backslashes = 0usize;
-    while cursor > 0 {
-        cursor -= 1;
-        if bytes[cursor] != b'\\' {
-            break;
-        }
-        backslashes += 1;
-    }
-
-    backslashes % 2 == 1
-}
-
-fn source_text_contains_shell_quoting_literals(text: &str) -> bool {
-    if text.contains(['"', '\'']) {
-        return true;
-    }
-
-    let chars = text.chars().collect::<Vec<_>>();
-    let mut index = 0usize;
-    while index < chars.len() {
-        if chars[index] != '\\' {
-            index += 1;
-            continue;
-        }
-
-        let mut end = index + 1;
-        while end < chars.len() && chars[end] == '\\' {
-            end += 1;
-        }
-        if chars.get(end).is_some_and(|next| {
-            matches!(next, '"' | '\'') || (next.is_whitespace() && !matches!(next, '\n' | '\r'))
-        }) {
-            return true;
-        }
-
-        index = end;
-    }
-
-    false
-}
-
-fn shell_quoting_segment_span(word: &Word, text: &str, start: usize, end: usize) -> Option<Span> {
-    let segment = &text[start..end];
-    if !source_text_contains_shell_quoting_literals(segment) {
-        return None;
-    }
-
-    let trimmed_start = if let Some(anchor) = first_escape_anchor(segment) {
-        segment[..anchor]
-            .rfind('\'')
-            .map_or(start, |quote| start + quote + 1)
-    } else {
-        start
-    };
-
-    Some(Span::from_positions(
-        word.span.start.advanced_by(&text[..trimmed_start]),
-        word.span.start.advanced_by(&text[..end]),
-    ))
-}
-
-fn first_shell_quoting_anchor(text: &str) -> Option<usize> {
-    let chars = text.char_indices().collect::<Vec<_>>();
-    for (index, (offset, ch)) in chars.iter().copied().enumerate() {
-        if matches!(ch, '"' | '\'') {
-            return Some(offset);
-        }
-        if ch != '\\' {
-            continue;
-        }
-        if let Some((_, next)) = chars.get(index + 1).copied()
-            && (matches!(next, '"' | '\'') || next.is_whitespace())
-        {
-            return Some(offset);
-        }
-    }
-
-    None
-}
-
-fn first_escape_anchor(text: &str) -> Option<usize> {
-    let chars = text.char_indices().collect::<Vec<_>>();
-    for (index, (offset, ch)) in chars.iter().copied().enumerate() {
-        if ch != '\\' {
-            continue;
-        }
-        if let Some((_, next)) = chars.get(index + 1).copied()
-            && (matches!(next, '"' | '\'') || next.is_whitespace())
-        {
-            return Some(offset);
-        }
-    }
-
-    first_shell_quoting_anchor(text)
+    word_shell_quoting_literal_run_span_in_source(binding.word, checker.source())
+        .unwrap_or(binding.word.span)
 }
 
 fn expansion_span_is_plain_reference(
